// pokebooook.client/src/types/PokemonTypeEnum.ts
type PokemonTypeEnum = {
    typeId: number;
    imageId: number;
    name: string;
    toGround: number;
    toGrass: number;
    toFire: number;
    toWater: number;
    toElectric: number;
    toPsychic: number;
    toFighting: number;
    toPoison: number;
    toFlying: number;
    toBug: number;
    toRock: number;
    toGhost: number;
    toDark: number;
    toDragon: number;
    toSteel: number;
    toFairy: number;
    toIce: number;
    toNormal: number;
};

export default PokemonTypeEnum;